requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
django-filter>=24.2
djangorestframework-simplejwt>=5.3
celery>=5.4
redis>=5.0
